# perf(prover): reduce synchronization allocs for FFT